FIX: In SendMail, now try to remove last error file only if exists.

+++ src/main/java/fr/ens/biologie/genomique/aozan/aozan3/SendMail.java
@@ -160,7 +160,9 @@ public class SendMail {
       this.logger.warn("Error while sending mail: " + mex.getMessage());
     }
 
-    if (!error && this.lastErrorFile != null) {
+    if (!error
+        && this.lastErrorFile != null
+        && Files.isRegularFile(this.lastErrorFile)) {
 
       try {
         Files.delete(this.lastErrorFile);
